requirements.txt
torch
torchvision
torchaudio
pandas
numpy
scikit-learn
matplotlib
seaborn
kagglehub
pillow
flask
joblib